+++ mlnn/engine.py
@@ -20,7 +20,7 @@ class MLNNEngine:
         self.check_array_equal = False
 
         if 'inner_loss' not in mlnn_params:
-            from activation import ReLU
+            from mlnn.activation import ReLU
             self.inner_loss = ReLU(1)
         if 'outer_loss' not in mlnn_params:
             self.outer_loss = self.inner_loss
